model: split in separate files
This should increase readability and discoverability.

Again, for blog post https://github.com/ooni/probe-engine/issues/746
