+++ model/README.md
@@ -1,0 +1,3 @@
+# Package github.com/ooni/probe-engine/model
+
+Shared data structures and interfaces.
