+++ model/service.go
@@ -1,0 +1,17 @@
+package model
+
+// Service describes a backend service.
+//
+// The fields of this struct have the meaning described in v2.0.0 of the OONI
+// bouncer specification defined by
+// https://github.com/ooni/spec/blob/master/backends/bk-004-bouncer.md.
+type Service struct {
+	// Address is the address of the server.
+	Address string `json:"address"`
+
+	// Type is the type of the service.
+	Type string `json:"type"`
+
+	// Front is the front to use with "cloudfront" type entries.
+	Front string `json:"front,omitempty"`
+}
